Add Mat implement(part)

+++ demo/test_core.cpp
@@ -575,7 +575,7 @@ int test_Scalar()
 
 int test_Mat()
 {
-	cv::Mat mat = cv::Mat(1, 111, CV_8UC1);
+	cv::Mat mat = cv::Mat(1, 111, CV_32FC1);
 	int stride = mat.step;
 	cv::Mat mat1_1 = cv::imread("E:/GitCode/OpenCV_Test/test_images/lena.png", 1);
 	if (!mat1_1.data) {
